dont run db:reset on production ENV

+++ lib/tasks/sample_data.rake
@@ -1,7 +1,7 @@
 namespace :db do
   desc "Fill database with sample data"
   task populate: :environment do
-    Rake::Task["db:reset"].invoke
+    Rake::Task["db:reset"].invoke if Rails.env.development?
     make_users
     make_microposts
     make_relationships
